jbcli/requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile
#
argh==0.26.2              # via watchdog
backports.csv==1.0.7      # via tablib
backports.ssl-match-hostname==3.7.0.1  # via docker, docker-compose
boto3==1.9.212
botocore==1.12.212
cached-property==1.5.1    # via docker-compose
certifi==2019.3.9
click==6.7
colorama==0.3.9           # via docker-compose
defusedxml==0.6.0         # via odfpy
docker-compose==1.12.0
docker-pycreds==0.4.0     # via docker
docker==2.2.1
dockerpty==0.4.1          # via docker-compose
docopt==0.6.2             # via docker-compose
docutils==0.15.2          # via botocore
elasticsearch-dsl==6.3.1
elasticsearch==6.3.1
enum34==1.1.6             # via docker-compose
et-xmlfile==1.0.1         # via openpyxl
functools32==3.2.3.post2  # via jsonschema
futures==3.3.0            # via s3transfer
humanize==0.5.1
ipaddress==1.0.22         # via docker, docker-compose, elasticsearch-dsl
jdcal==1.4.1              # via openpyxl
jmespath==0.9.4           # via boto3, botocore
jsonschema==2.6.0         # via docker-compose
odfpy==1.4.0              # via tablib
openpyxl==2.6.3           # via tablib
pathtools==0.1.2          # via watchdog
prompt_toolkit==1.0.14    # via pyinquirer
pygments==2.4.2           # via pyinquirer
pyinquirer==1.0.3
python-dateutil==2.8.0    # via botocore, elasticsearch-dsl
pyyaml==3.13              # via docker-compose, tablib, watchdog
regex==2019.8.19          # via pyinquirer
requests==2.11.1          # via docker, docker-compose
s3transfer==0.2.1         # via boto3
six==1.12.0               # via docker, docker-compose, docker-pycreds, dockerpty, elasticsearch-dsl, prompt-toolkit, python-dateutil, websocket-client
tablib==0.13.0
tabulate==0.8.3
texttable==0.8.8          # via docker-compose
toml==0.10.0
urllib3==1.24.3           # via botocore, elasticsearch
watchdog==0.8.3
wcwidth==0.1.7            # via prompt-toolkit
websocket-client==0.56.0  # via docker, docker-compose
xlrd==1.2.0               # via tablib
xlwt==1.3.0               # via tablib
